Change PurityPercentage field to double

+++ ConsoleApp/Product.cs
@@ -9,6 +9,6 @@ public enum AddLevel {
 public class Product {
     private string Name { get; set; } = null!;
     private int PricePerPound { get; set; }
-    private int PurityPercentage { get; set; }
+    private double PurityPercentage { get; set; }
     private AddLevel AddictivenessLevel { get; set; }
 }
